requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=./requirements.txt ./requirements.in
#
amqp==2.6.1
    # via kombu
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via
    #   flake8-bugbear
    #   jsonschema
    #   pytest
billiard==3.6.4.0
    # via celery
black==21.6b0
    # via -r ./requirements.in
bleach==3.3.1
    # via django-sql-dashboard
bumpver==2021.1113
    # via -r ./requirements.in
celery==4.4.7
    # via -r ./requirements.in
certifi==2021.5.30
    # via
    #   requests
    #   sentry-sdk
chardet==3.0.4
    # via requests
click==7.1.2
    # via
    #   black
    #   bumpver
    #   django-click
colorama==0.4.4
    # via
    #   bumpver
    #   rich
commonmark==0.9.1
    # via rich
django==3.2.5
    # via
    #   -r ./requirements.in
    #   django-extensions
    #   django-filter
    #   django-improved-user
    #   django-sql-dashboard
    #   djangorestframework
    #   drf-spectacular
    #   model-bakery
django-click==2.2.0
    # via -r ./requirements.in
django-constance[redis]==2.8.0
    # via -r ./requirements.in
django-environ==0.4.5
    # via -r ./requirements.in
django-extensions==3.1.3
    # via -r ./requirements.in
django-filter==2.4.0
    # via -r ./requirements.in
django-improved-user==1.0.1
    # via -r ./requirements.in
django-sql-dashboard==1.0.1
    # via -r ./requirements.in
django-test-plus==2.2.0
    # via -r ./requirements.in
djangorestframework==3.12.4
    # via
    #   -r ./requirements.in
    #   drf-spectacular
djhtml==1.4.9
    # via -r ./requirements.in
drf-spectacular==0.17.2
    # via -r ./requirements.in
factory-boy==3.2.0
    # via -r ./requirements.in
faker==8.10.2
    # via
    #   -r ./requirements.in
    #   factory-boy
flake8==3.9.2
    # via
    #   -r ./requirements.in
    #   flake8-bugbear
    #   flake8-docstrings
    #   flake8-polyfill
    #   pep8-naming
flake8-bugbear==21.4.3
    # via -r ./requirements.in
flake8-docstrings==1.6.0
    # via -r ./requirements.in
flake8-polyfill==1.0.2
    # via pep8-naming
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
jsonschema==3.2.0
    # via drf-spectacular
kombu==4.6.11
    # via celery
lexid==2021.1006
    # via bumpver
markdown==3.3.4
    # via django-sql-dashboard
mccabe==0.6.1
    # via flake8
model-bakery==1.3.2
    # via -r ./requirements.in
mypy-extensions==0.4.3
    # via black
packaging==21.0
    # via
    #   bleach
    #   pytest
pathlib2==2.3.6
    # via bumpver
pathspec==0.9.0
    # via black
pep8-naming==0.12.0
    # via -r ./requirements.in
pluggy==0.13.1
    # via pytest
psycopg2-binary==2.9.1
    # via -r ./requirements.in
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.3.1
    # via flake8
pygments==2.9.0
    # via rich
pyparsing==2.4.7
    # via packaging
pyrsistent==0.18.0
    # via jsonschema
pytest==6.2.4
    # via
    #   -r ./requirements.in
    #   pytest-django
    #   pytest-lazy-fixture
pytest-django==4.4.0
    # via -r ./requirements.in
pytest-lazy-fixture==0.6.3
    # via -r ./requirements.in
python-dateutil==2.8.2
    # via
    #   -r ./requirements.in
    #   faker
    #   time-machine
pytz==2018.5
    # via
    #   -r ./requirements.in
    #   celery
    #   django
pyyaml==5.4.1
    # via drf-spectacular
redis==3.5.3
    # via
    #   -r ./requirements.in
    #   django-constance
regex==2021.7.6
    # via black
requests==2.23.0
    # via -r ./requirements.in
rich==10.6.0
    # via -r ./requirements.in
sentry-sdk==1.4.3
    # via -r ./requirements.in
six==1.16.0
    # via
    #   bleach
    #   jsonschema
    #   pathlib2
    #   python-dateutil
snowballstemmer==2.1.0
    # via pydocstyle
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
time-machine==2.3.1
    # via -r ./requirements.in
toml==0.10.2
    # via
    #   black
    #   bumpver
    #   pytest
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.25.11
    # via
    #   requests
    #   sentry-sdk
vine==1.3.0
    # via
    #   amqp
    #   celery
webencodings==0.5.1
    # via bleach
whitenoise==5.3.0
    # via -r ./requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
